day-13-deepseek-py/requirements.txt
# Core
pandas>=1.3.0
numpy>=1.21.0
sqlalchemy>=1.4.0
python-dotenv>=0.19.0

# Big Data
pyspark>=3.3.0

# Cloud
boto3>=1.24.0
google-cloud-storage>=2.0.0

# Data Validation
great-expectations>=0.15.0

# Logging
structlog>=21.1.0